Convert library entry point to Typescript.
Modify configuration files to compile this new file and include it with integration testing.
